requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.1.7
django-debug-toolbar==3.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
numpy==1.20.2
pandas==1.2.3
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
urllib3==1.26.4
w3lib==1.22.0
webdriver-manager==3.3.0
zope.interface==5.3.0
